+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Main
 
+* [5184](https://github.com/grafana/loki/pull/5184) **ssncferreira** Deprecate `querier.split-queries-by-interval` in favor of `limits.split-queries-by-interval`.
 * [5170](https://github.com/grafana/loki/pull/5170) **chaudum** Fix deadlock in Promtail caused when targets got removed from a target group by the discovery manager.
 * [5163](https://github.com/grafana/loki/pull/5163) **chaudum** Fix regression in fluentd plugin introduced with #5107 that caused `NoMethodError` when parsing non-string values of log lines.
 * [5144](https://github.com/grafana/loki/pull/5144) **dannykopping** Ruler: fix remote write basic auth credentials.